refactor: move ChartAreaFlag
