+++ src/libecalc/core/models/compressor/sampled/compressor_model_sampled.py
@@ -12,6 +12,7 @@ from libecalc.common.list.adjustment import transform_linear
 from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.logger import logger
 from libecalc.common.units import Unit
+from libecalc.core.models.chart.chart_area_flag import ChartAreaFlag
 from libecalc.core.models.compressor.base import CompressorModel
 from libecalc.core.models.compressor.sampled.compressor_model_sampled_1d import (
     CompressorModelSampled1D,
@@ -42,7 +43,6 @@ from libecalc.core.models.results.compressor import (
 )
 from libecalc.domain.stream_conditions import StreamConditions
 from libecalc.dto.models.compressor import CompressorSampled
-from libecalc.dto.types import ChartAreaFlag
 
 
 class CompressorModelSampled(CompressorModel):
