+++ src/libecalc/dto/types.py
@@ -129,17 +129,3 @@ class FuelType(EcalcBaseModel):
                 )
 
         return user_defined_category
-
-
-class ChartAreaFlag(str, Enum):
-    INTERNAL_POINT = "INTERNAL_POINT"
-    BELOW_MINIMUM_FLOW_RATE = "BELOW_MINIMUM_FLOW_RATE"
-    BELOW_MINIMUM_HEAD = "BELOW_MINIMUM_HEAD"
-    BELOW_MINIMUM_SPEED = "BELOW_MINIMUM_SPEED"
-    ABOVE_MAXIMUM_FLOW_RATE = "ABOVE_MAXIMUM_FLOW_RATE"
-    ABOVE_MAXIMUM_HEAD = "ABOVE_MAXIMUM_HEAD"
-    ABOVE_MAXIMUM_SPEED = "ABOVE_MAXIMUM_SPEED"
-    BELOW_MINIMUM_SPEED_AND_BELOW_MINIMUM_FLOW_RATE = "BELOW_MINIMUM_SPEED_AND_BELOW_MINIMUM_FLOW_RATE"
-    BELOW_MINIMUM_SPEED_AND_ABOVE_MAXIMUM_FLOW_RATE = "BELOW_MINIMUM_SPEED_AND_ABOVE_MAXIMUM_FLOW_RATE"
-    NOT_CALCULATED = "NOT_CALCULATED"
-    NO_FLOW_RATE = "NO_FLOW_RATE"
