+++ src/tests/libecalc/core/models/compressor_modelling/test_compressor_chart.py
@@ -4,11 +4,11 @@ from pytest import approx
 
 from libecalc import dto
 from libecalc.common.errors.exceptions import IllegalStateException
+from libecalc.core.models.chart.chart_area_flag import ChartAreaFlag
 from libecalc.core.models.compressor.train.chart import (
     SingleSpeedCompressorChart,
     VariableSpeedCompressorChart,
 )
-from libecalc.dto.types import ChartAreaFlag
 
 
 @pytest.fixture
