+++ src/tests/libecalc/dto/results/model/test_extend.py
@@ -1,13 +1,13 @@
 import pytest
 
 from libecalc.common.units import Unit
+from libecalc.core.models.chart.chart_area_flag import ChartAreaFlag
 from libecalc.core.models.results.base import EnergyFunctionResult
 from libecalc.core.models.results.compressor import (
     CompressorStageResult,
     CompressorStreamCondition,
     CompressorTrainResult,
 )
-from libecalc.dto.types import ChartAreaFlag
 
 
 @pytest.fixture
